==> notebooks/03.04 - Data - Calendar Data - Compile.ipynb <==
# 03.04 - Calendar Data

## Imports & setup

import pathlib
import datetime
import dateutil
from os import PathLike
from typing import Union

#import simplegeneric
import pandas as pd
import numpy as np
from astral import Astral

import matplotlib.pyplot as plt
plt.style.use('grayscale')
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
import palettable
%matplotlib inline


PROJECT_DIR = pathlib.Path.cwd().parent.resolve()
IMPUTED_DATA_DIR_DEMAND = PROJECT_DIR / 'data' / '03-imputed' / 'demand'
CALCULATED_FEATURES_DATA_DIR = PROJECT_DIR / 'data' / '03-calculated-features' / 'calendar'

## Load

demand_df = pd.read_csv(IMPUTED_DATA_DIR_DEMAND / 'demand.csv', index_col=0, parse_dates=True,
                       date_parser=dateutil.parser.parse)
#demand_df.index.tz_localize(None)
demand_df.info()

features_df = demand_df.copy(deep=True)
features_df['hour_of_day'] = features_df.index.hour
features_df['year'] = features_df.index.year
features_df['month'] = features_df.index.month
features_df['day_of_week'] = features_df.index.dayofweek
features_df['day_of_year'] = features_df.index.dayofyear
features_df['week_of_year'] = features_df.index.weekofyear
features_df['quarter'] = features_df.index.quarter
features_df.drop(columns=['ont_demand'], inplace=True)

features_df.head()

import holidays


hols = holidays.Canada(state='ON') # default is ontario Holidays
print(features_df.loc['2018-01-01'].index.date[0] in hols)
print(features_df.loc['2018-12-27'].index.date[0] in hols)

features_df['stat_hol'] = pd.Series(features_df.index.date).apply(lambda x: x in hols).values
features_df.head()

features_df.tail()

from astral import Astral
a = Astral()
city_name='Toronto'
city = a[city_name]
#city.latitude
sun = city.sun(date=datetime.date(2019, 7, 2), local=True)
print(sun['sunrise'])
print(sun['sunset'])
print(type(sun['sunrise']))

print(features_df.loc['2018-01-01'].index[0])
print(features_df.loc['2018-12-27'].index[0])

features_df.head()

def get_daylight_hours(row, city):
    sun = city.sun(date=row.name, local=True)
    sunrise = sun['sunrise'].replace(tzinfo=None) ; sunset = sun['sunset'].replace(tzinfo=None)
    bool_val = (row.name > sunrise) & (row.name < sunset)
    return bool_val


a = Astral()
city = a['Toronto']
features_df['day_light_hours'] = features_df.apply(get_daylight_hours, city=city, axis=1)
features_df.head()

features_df.tail()

features_df.to_csv(CALCULATED_FEATURES_DATA_DIR / 'calendar.csv')

features_df.info()
