==> notebooks/03.05 - Data - Merge Demand, Weather, Calendar Data.ipynb <==
# 03.05 - Merge Demand, Weather, Calendar Data

## Imports & setup

import pathlib
import datetime
import dateutil
from os import PathLike
from typing import Union

#import simplegeneric
import pandas as pd
import numpy as np
from astral import Astral

import matplotlib.pyplot as plt
plt.style.use('grayscale')
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
import palettable
%matplotlib inline


PROJECT_DIR = pathlib.Path.cwd().parent.resolve()

IMPUTED_DATA_DIR_DEMAND = PROJECT_DIR / 'data' / '03-imputed' / 'demand'
INTERIM_DATA_DIR_DEMAND = PROJECT_DIR / 'data' / '04-interim' / 'demand'
CALCULATED_FEATURES_DATA_DIR_CALENDAR = PROJECT_DIR / 'data' / '03-calculated-features' / 'calendar'
IMPUTED_DATA_DIR_WEATHER = PROJECT_DIR / 'data' / '03-imputed' / 'weather-toronto'
CLEAN_DATA_DIR = PROJECT_DIR / 'data' / '05-clean'

## Join Calendar features and Imputed Demand

cal_df = pd.read_csv(CALCULATED_FEATURES_DATA_DIR_CALENDAR / 'calendar.csv', index_col=0, parse_dates=True,
                       date_parser=dateutil.parser.parse)
cal_df.info()

power_df = pd.read_csv(IMPUTED_DATA_DIR_DEMAND / 'demand.csv', index_col=0, parse_dates=True,
                      date_parser=dateutil.parser.parse)
power_df.rename(columns={'ont_demand': 'hourly_demand'}, inplace=True)

power_df.info()

feat_df = pd.concat([cal_df, power_df], axis=1, sort=True)
feat_df.head()

feat_df.info()

feat_df.to_csv(INTERIM_DATA_DIR_DEMAND / 'demand.csv')

## Join Demand, Demand Features, Calendar features, Weather Features

int_df = pd.read_csv(INTERIM_DATA_DIR_DEMAND / 'demand.csv', index_col=0, parse_dates=True,
                       date_parser=dateutil.parser.parse)
int_df.rename(columns ={'ont_demand': 'hourly_demand'}, inplace=True)
int_df.info()

int_df.head()

int_df.tail()

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,
          'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,
          'hmdx': np.float64, 'wind_chill': np.float64, 'weather': np.object}
weather_df = pd.read_csv(IMPUTED_DATA_DIR_WEATHER / "weather_toronto.csv", parse_dates=True,
                 dtype=dtypes, index_col=0, infer_datetime_format=True)

weather_df.info()

weather_df.head()

weather_df.tail()

clean_df = pd.concat([weather_df, int_df], axis=1, sort=True)
clean_df.head()

clean_df.tail()

clean_df.info()

# Create a daily_peak column
# Peak Daily demand
clean_df = clean_df.join(feat_df.groupby(by=['year', 'day_of_year'])['hourly_demand'].max(),
                       on=['year', 'day_of_year'], rsuffix='_peak_in_day_in_year')
clean_df.rename(columns={'hourly_demand_peak_in_day_in_year': 'daily_peak'}, inplace=True)

clean_df.to_csv(CLEAN_DATA_DIR / 'oversized.csv')

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,
          'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,
          'hmdx': np.float64, 'wind_chill': np.float64, 'weather': np.object,
         'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,
         'day_of_week': np.float64, 'day_of_year': np.float64, 'week_of_year': np.float64,
          'quarter': np.float64, 'stat_hol': np.object, 'day_light_hours': np.object,
          'hourly_demand': np.float64, 'daily_peak': np.float64, 'weekly_peak': np.float64,
          'monthly_peak': np.float64, 'quarterly_peak': np.float64, 'annual_peak': np.float64,
          'hour_as_pct_of_daily_peak': np.float64, 'daily_peak_as_pct_of_weekly_peak': np.float64,
         'monthly_peak_as_pct_of_annual_peak': np.float64,
          'quarterly_peak_as_pct_of_annual_peak': np.float64, 'daily_peak_as_pct_of_annual_peak': np.float64}

oversized_df = pd.read_csv(CLEAN_DATA_DIR / "oversized.csv", parse_dates=True,
                 dtype=dtypes, index_col=0, infer_datetime_format=True)

oversized_df.info()

clean_df = oversized_df.dropna(subset=['hourly_demand'])

clean_df.to_csv(CLEAN_DATA_DIR / "clean.csv")

clean_df.info()

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,
          'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,
          'hmdx': np.float64, 'wind_chill': np.float64, 'weather': np.object,
         'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,
         'day_of_week': np.float64, 'day_of_year': np.float64, 'week_of_year': np.float64,
          'quarter': np.float64, 'stat_hol': np.object, 'day_light_hours': np.object,
          'hourly_demand': np.float64, 'daily_peak': np.float64, 'weekly_peak': np.float64,
          'monthly_peak': np.float64, 'quarterly_peak': np.float64, 'annual_peak': np.float64,
          'hour_as_pct_of_daily_peak': np.float64, 'daily_peak_as_pct_of_weekly_peak': np.float64,
         'monthly_peak_as_pct_of_annual_peak': np.float64,
          'quarterly_peak_as_pct_of_annual_peak': np.float64, 'daily_peak_as_pct_of_annual_peak': np.float64}

df = pd.read_csv(CLEAN_DATA_DIR / "clean.csv", parse_dates=True,
                 dtype=dtypes, index_col=0, infer_datetime_format=True)

df.info()

clean_cut_df = clean_df.loc['1994': '2018']
clean_cut_df.to_csv(CLEAN_DATA_DIR / "clean-cut.csv")

dtypes = {'temp': np.float64, 'dew_point_temp':np.float64, 'rel_hum':np.float64,
          'wind_speed': np.float64, 'visibility': np.float64, 'press': np.float64,
          'hmdx': np.float64, 'wind_chill': np.float64, 'weather': np.object,
         'hour_of_day': np.float64, 'year': np.float64, 'month': np.float64,
         'day_of_week': np.float64, 'day_of_year': np.float64, 'week_of_year': np.float64,
          'quarter': np.float64, 'stat_hol': np.object, 'day_light_hours': np.object,
          'hourly_demand': np.float64, 'daily_peak': np.float64, 'weekly_peak': np.float64,
          'monthly_peak': np.float64, 'quarterly_peak': np.float64, 'annual_peak': np.float64,
          'hour_as_pct_of_daily_peak': np.float64, 'daily_peak_as_pct_of_weekly_peak': np.float64,
         'monthly_peak_as_pct_of_annual_peak': np.float64,
          'quarterly_peak_as_pct_of_annual_peak': np.float64, 'daily_peak_as_pct_of_annual_peak': np.float64}

clean_cut_df = pd.read_csv(CLEAN_DATA_DIR / "clean-cut.csv", parse_dates=True,
                 dtype=dtypes, index_col=0, infer_datetime_format=True)

clean_cut_df.info()

clean_cut_df.head()

clean_cut_df.tail()
