==> notebooks/05.02 - Modeling Setup - Scoring.ipynb <==
# 05.02 - Modeling Setup - Custom Scoring

 + Imports & Setup
 + Load Test Data
 + Custom Scoring Metric

## Imports & setup

import pathlib
from datetime import datetime
import math
import sys

import pandas as pd
import numpy as np

import matplotlib as mpl
import matplotlib.pyplot as plt
plt.style.use('grayscale')
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates

from sklearn.metrics import mean_absolute_error

sys.path.append("..")
from src.models.models import SetTempAsPower, SK_Prophet

%matplotlib inline

PROJECT_DIR = pathlib.Path.cwd().parent.resolve()
CLEAN_DATA_DIR = PROJECT_DIR / 'data' /'05-clean'

pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

## Load Test Data

df = pd.read_csv(CLEAN_DATA_DIR / 'clean-cut.csv', parse_dates=True, index_col=0)
df = df.loc['1994': '1995']
df = df.resample('D').max()
# Just select a reasonable subset of data to test the model wrappers
df = df[['temp', 'dew_point_temp', 'week_of_year', 'daily_peak']]
df.rename(columns={'temp': 'temp_max'}, inplace=True)

y = df.pop('daily_peak')
X = df

X.head()

y.head()

y.tail()

##  Custom Scoring Metric

def bound_precision(y_actual: pd.Series, y_predicted: pd.Series, n_to_check=5):
    """
    Accepts two pandas series, and an integer n_to_check
    Series are:
    + actual values
    + predicted values
    Sorts each series by value from high to low, and cuts off each series at n_to_check
    Determines how many hits - ie how many of the indices in the actual series are in the predicted series indices
    Returns number of hits divided by n_to_check    
    """
    y_act = y_actual.copy(deep=True)
    y_pred = y_predicted.copy(deep=True)
    y_act.reset_index(drop=True, inplace=True)
    y_pred.reset_index(drop=True, inplace=True)

    act_dates =set( y_act.sort_values(ascending=False).head(n_to_check).index)
    pred_dates = set(y_pred.sort_values(ascending=False).head(n_to_check).index)
    bound_precision =  len(act_dates.intersection(pred_dates))/ n_to_check
    return bound_precision

y_act = pd.Series([ 11,12,13,14, 15,16,17, 11, 12], index = pd.date_range(start='2019-01-01', periods=9))
y_pred = pd.Series([18,11,13,14, 16,15,15, 14, 11], index = pd.date_range(start='2019-03-20', periods=9))
b_prec = bound_precision(y_act, y_pred, n_to_check=3)
b_prec

X_m = X.copy(deep=True)
y_m = y.copy(deep=True)
X_train = X_m['1994'] ; y_train = y_m['1994']
X_test = X_m['1995'] ; y_test = y_m['1995']

set_temp_as_power = SetTempAsPower(col='temp_max')
set_temp_as_power.fit(X_train, y_train)
preds = set_temp_as_power.predict(X_test)
print(preds)
print()
print(mean_absolute_error(y_test, preds))
print()
print(bound_precision(y_test, preds))

sk_prophet = SK_Prophet(regressors={'temp_max':()})
sk_prophet.fit(X_train, y_train)
preds = sk_prophet.predict(X_test)
print(preds)
print()
print(mean_absolute_error(y_test, preds))
print()
print(bound_precision(y_test, preds))
