==> notebooks/07.08 - Modeling - SARIMAX AutoARIMA.ipynb <==
# 07.08 - SARIMAX AutoARIMA - 5 Years Training Data

## Imports & setup

import pathlib
import warnings
from datetime import datetime
import sys
import pickle
import joblib
import gc

import pandas as pd
import numpy as np

# Plotting
import matplotlib as mpl
import matplotlib.pyplot as plt
from pandas.plotting import register_matplotlib_converters
register_matplotlib_converters()
from matplotlib.dates import DateFormatter
import matplotlib.dates as mdates
import palettable
import seaborn as sns

# Imports
sys.path.append("..")
from src.utils.utils import (AnnualTimeSeriesSplit,
                             bound_precision,
                             run_cross_val,
                             run_data_split_cross_val,
                             save_run_results)
from src.features.features import CyclicalToCycle
from src.models.models import SK_SARIMAX
from src.visualization.visualize import (plot_prediction,
                                         plot_joint_plot,
                                         residual_plots,
                                         print_residual_stats)
#b # Packages
from sklearn.pipeline import Pipeline
from skoot.feature_selection import FeatureFilter
from skoot.preprocessing import SelectiveRobustScaler
from sklearn.metrics import mean_absolute_error
from scipy.stats import norm
from statsmodels.graphics.gofplots import qqplot
from pandas.plotting import autocorrelation_plot
from statsmodels.graphics.tsaplots import plot_acf
import statsmodels.api as sm
import pmdarima as pm



inline_rc = dict(mpl.rcParams)


# Display
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
figsize=(15,7)
warnings.filterwarnings(action='ignore')
%matplotlib inline

# Data
PROJECT_DIR = pathlib.Path.cwd().parent.resolve()
CLEAN_DATA_DIR = PROJECT_DIR / 'data' / '05-clean'

### Auto ARIMA

we can use AutoArima to look for the best values of p, d, q, P, D, and Q

At this point, we will throw in all the exogenous variables without any pre-processing. We can fine tune this later, after we have decided on the ARIMA parameters

Auto Arima can only handle 3 years at a time on my computer before crashing with a memory error. Therefore, we will run x batches of 3 years, and review the results in order to select the ARIMA parameters

## Load Daily Data & Inspect

df = pd.read_csv(CLEAN_DATA_DIR / 'clean-features.csv', parse_dates=True, index_col=0)

X = df.loc['2004': '2008']
y = X.pop('daily_peak')

X.head()

y.tail()

## Run Auto ARIMA

# Set up Feature Transformers
cycle0 = CyclicalToCycle('hmdxx_max_hour', 24)
cycle1 = CyclicalToCycle('sun_rise', 24)
cycle2 = CyclicalToCycle('sun_set', 24)
cycle3 = CyclicalToCycle('day_of_week', 5)
cycle4 = CyclicalToCycle('week_of_year', 20)

robust_scaler_cols = ['hmdxx_min', 'hmdxx_max', 'hmdxx_median-1', 'temp_min', 'temp_max',
                      'dew_point_temp_max', 'visibility_mean']

rscaler = SelectiveRobustScaler(cols=robust_scaler_cols, trans_col_name=robust_scaler_cols)


cycle0.fit(X, y)
X_t = cycle0.transform(X)

cycle1.fit(X_t, y)
X_t = cycle1.transform(X_t)

cycle2.fit(X_t, y)
X_t = cycle2.transform(X_t)

cycle3.fit(X_t, y)
X_t = cycle3.transform(X_t)

cycle4.fit(X_t, y)
X_t = cycle4.transform(X_t)

rscaler.fit(X_t, y)
X_t = rscaler.transform(X_t)

# Stepwise Fit
model = pm.auto_arima(y,
                      seasonal=True,
                      m=96,
                      trace=True,
                      trend='c',
                      exog=X_t,
                      error_action='ignore',
                      suppress_warnings=True)

model.summary()

model.get_params()

X_t.head()

X_t.tail()

Fit ARIMA: order=(2, 1, 2) seasonal_order=(1, 0, 1, 96); AIC=8275.806, BIC=8309.179, Fit time=235.733 seconds
Fit ARIMA: order=(0, 1, 0) seasonal_order=(0, 0, 0, 96); AIC=8330.977, BIC=8339.321, Fit time=0.020 seconds
Fit ARIMA: order=(1, 1, 0) seasonal_order=(1, 0, 0, 96); AIC=8331.396, BIC=8348.083, Fit time=22.633 seconds
Fit ARIMA: order=(0, 1, 1) seasonal_order=(0, 0, 1, 96); AIC=8330.911, BIC=8347.598, Fit time=15.204 seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(0, 0, 1, 96); AIC=8276.142, BIC=8305.344, Fit time=108.742 seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(2, 0, 1, 96); AIC=8277.849, BIC=8315.394, Fit time=798.965 seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(1, 0, 0, 96); AIC=8276.287, BIC=8305.489, Fit time=86.259 seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(1, 0, 2, 96); AIC=nan, BIC=nan, Fit time=nan seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(0, 0, 0, 96); AIC=8278.078, BIC=8303.108, Fit time=0.856 seconds
Fit ARIMA: order=(2, 1, 2) seasonal_order=(2, 0, 2, 96); AIC=nan, BIC=nan, Fit time=nan seconds
Fit ARIMA: order=(1, 1, 2) seasonal_order=(1, 0, 1, 96); AIC=8277.909, BIC=8307.111, Fit time=141.145 seconds
Fit ARIMA: order=(3, 1, 2) seasonal_order=(1, 0, 1, 96); AIC=8277.285, BIC=8314.831, Fit time=168.277 seconds
Fit ARIMA: order=(2, 1, 1) seasonal_order=(1, 0, 1, 96); AIC=8277.052, BIC=8306.254, Fit time=140.057 seconds
Fit ARIMA: order=(2, 1, 3) seasonal_order=(1, 0, 1, 96); AIC=8277.552, BIC=8315.098, Fit time=167.539 seconds
Fit ARIMA: order=(1, 1, 1) seasonal_order=(1, 0, 1, 96); AIC=8283.910, BIC=8308.940, Fit time=129.563 seconds
Fit ARIMA: order=(3, 1, 3) seasonal_order=(1, 0, 1, 96); AIC=8279.325, BIC=8321.042, Fit time=216.832 seconds
Total fit time: 2232.217 seconds

There is not too much difference between the best models, so we will choose a variety from these to try out in our cross validation scheme:

+ SARIMAX I (2,1,2) (1,0,1,96)
+ SARIMAX II (2,1,2) (0,0,1,96)
+ SARIMAX III (2,1,2) (1,0,0,96)
+ SARIMAX IV (1,0,1) (1,0,0,96) - Made this one up to try as a reference
